requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dacite==1.8.1
dnspython==2.5.0
exceptiongroup==1.2.0
fastapi==0.109.0
fonttools==4.47.2
gunicorn==21.2.0
h11==0.14.0
htmlmin==0.1.12
idna==3.6
ImageHash==4.3.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.1
MarkupSafe==2.1.4
matplotlib==3.8.2
motor==3.3.2
multimethod==1.11
networkx==3.2.1
numba==0.58.1
numpy==1.25.2
packaging==23.2
pandas==2.2.0
patsy==0.5.6
phik==0.12.4
pillow==10.2.0
pydantic==2.6.0
pydantic_core==2.16.1
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
statsmodels==0.14.1
tangled-up-in-unicode==0.2.0
tqdm==4.66.1
typeguard==4.1.5
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
visions==0.7.5
wordcloud==1.9.3
ydata-profiling==4.6.4
